[Agent] Extract default component injection
